# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API Documentation",
        "contact": {
            "name": "API Support",
            "email": "your.email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a list of all clients",
                "operationId": "4301a2cf786f38221ea35bc013867a24",
                "responses": {
                    "200": {
                        "description": "List of clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "operationId": "fe94ce0c3c351c97ff50833d2368655d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a specific client by ID",
                "operationId": "162a754c1f4e01a924233e3544d45681",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update an existing client",
                "operationId": "1155f52b3b1a0808702c7e71cd54ad65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete a client",
                "operationId": "2d01fd546ff33b908418ebee7436f935",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contractors": {
            "get": {
                "summary": "Retrieve all contractors",
                "operationId": "a2ba50b46942507b1a36791efbabc8bc",
                "responses": {
                    "200": {
                        "description": "List of contractors"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "Client ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Client name",
                        "type": "string"
                    },
                    "email": {
                        "description": "Client email address",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Client phone number",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Contractor": {
                "properties": {
                    "id": {
                        "description": "Contractor ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "contractor_name": {
                        "description": "Contractor name",
                        "type": "string"
                    },
                    "commission_percentage": {
                        "description": "Commission percentage",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Task": {
                "properties": {
                    "id": {
                        "description": "Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string"
                    },
                    "client_id": {
                        "description": "Client ID",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Task status",
                        "type": "string"
                    },
                    "mk": {
                        "description": "MK flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "mkcold": {
                        "description": "MK Cold flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "osv": {
                        "description": "OSV flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "osvEvak": {
                        "description": "OSV Evak flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "sh": {
                        "description": "SH flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "shobSgr": {
                        "description": "Shob SGR flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "shokolSr": {
                        "description": "Shokol SR flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "vent": {
                        "description": "Ventilation flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "klim": {
                        "description": "Climate flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "f0": {
                        "description": "F0 flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "z": {
                        "description": "Z flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "m": {
                        "description": "M flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "izol": {
                        "description": "Isolation flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "dtz": {
                        "description": "DTZ flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "mk_next": {
                        "description": "MK next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "mkcold_next": {
                        "description": "MK Cold next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "osv_next": {
                        "description": "OSV next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "osvEvak_next": {
                        "description": "OSV Evak next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "sh_next": {
                        "description": "SH next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "shobSgr_next": {
                        "description": "Shob SGR next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "shokolSr_next": {
                        "description": "Shokol SR next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "vent_next": {
                        "description": "Ventilation next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "klim_next": {
                        "description": "Climate next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "f0_next": {
                        "description": "F0 next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "z_next": {
                        "description": "Z next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "m_next": {
                        "description": "M next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "izol_next": {
                        "description": "Isolation next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "dtz_next": {
                        "description": "DTZ next flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "paid": {
                        "description": "Paid flag",
                        "type": "string",
                        "enum": [
                            "0",
                            "1"
                        ]
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Clients",
            "description": "Clients"
        }
    ]
}